// web/src/api/generated/models/CommentDetail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI ist der verbindliche Vertrag.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AuthorSummary } from './AuthorSummary';
import {
    AuthorSummaryFromJSON,
    AuthorSummaryFromJSONTyped,
    AuthorSummaryToJSON,
    AuthorSummaryToJSONTyped,
} from './AuthorSummary';

/**
 * 
 * @export
 * @interface CommentDetail
 */
export interface CommentDetail {
    /**
     * 
     * @type {AuthorSummary}
     * @memberof CommentDetail
     */
    author: AuthorSummary;
    /**
     * 
     * @type {string}
     * @memberof CommentDetail
     */
    authorId: string;
    /**
     * 
     * @type {string}
     * @memberof CommentDetail
     */
    body: string;
    /**
     * 
     * @type {Date}
     * @memberof CommentDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof CommentDetail
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof CommentDetail
     */
    spaceId: string;
    /**
     * 
     * @type {Date}
     * @memberof CommentDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof CommentDetail
     */
    version: number;
}

/**
 * Check if a given object implements the CommentDetail interface.
 */
export function instanceOfCommentDetail(value: object): value is CommentDetail {
    if (!('author' in value) || value['author'] === undefined) return false;
    if (!('authorId' in value) || value['authorId'] === undefined) return false;
    if (!('body' in value) || value['body'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('spaceId' in value) || value['spaceId'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    if (!('version' in value) || value['version'] === undefined) return false;
    return true;
}

export function CommentDetailFromJSON(json: any): CommentDetail {
    return CommentDetailFromJSONTyped(json, false);
}

export function CommentDetailFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommentDetail {
    if (json == null) {
        return json;
    }
    return {
        
        'author': AuthorSummaryFromJSON(json['author']),
        'authorId': json['authorId'],
        'body': json['body'],
        'createdAt': (new Date(json['createdAt'])),
        'id': json['id'],
        'spaceId': json['spaceId'],
        'updatedAt': (new Date(json['updatedAt'])),
        'version': json['version'],
    };
}

export function CommentDetailToJSON(json: any): CommentDetail {
    return CommentDetailToJSONTyped(json, false);
}

export function CommentDetailToJSONTyped(value?: CommentDetail | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'author': AuthorSummaryToJSON(value['author']),
        'authorId': value['authorId'],
        'body': value['body'],
        'createdAt': value['createdAt'].toISOString(),
        'id': value['id'],
        'spaceId': value['spaceId'],
        'updatedAt': value['updatedAt'].toISOString(),
        'version': value['version'],
    };
}